RAG-Proj/requirements.txt
pymupdf==1.26.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.36.2
streamlit==1.31.0
python-dotenv==1.0.0
requests==2.31.0
torch==2.2.0
